Updated WeightsCalc.java
Added main method which calls runCalc() the method which will run the
calculator.
Added a userInput() method to read from the command line.
Added sets & gets for instance attributes.
Added runCalc() stub which will eventually ask the user to input their
bar and weights and a then perform calculations on them until the user
enters nothing for a target.