Actually updated WeightsCalc to correctly calculate weights
calculateWeights method will correctly calculate weights now!  Still
only calculates to the nearest weight under the target though.  I may
fiddle with it to calculate the nearest weight achievable regardless of
under or over.

targetWeightString method developed.  Largely a copy and paste of
toString at the moment so they are an obvious candidate for a helper
method.

runCalc now asks the user to repeatedly input a target weight then
passes that input to calculateWeights and prints the calculated weights
to meet that target.  If the target is not achievable the user is
warned.